Made the Image from Backend to be rendered on frontend.
Suffered with the display of the image because I didn't know that I need to edit the vite.config.ts, which if you don't allow the path from there then it doesn't display anything.

Lesson learnt:
You need to allow also the paths when calling the api from the backend.